gitmark/requirements.txt
Django==1.6
south==0.8.4
djangorestframework==2.3.14